+++ aisdb/aisdb_sql/psql_createtable_static.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS ais_{0}_static (
-    mmsi INTEGER NOT NULL,
-    time INTEGER NOT NULL,
+    mmsi BIGINT NOT NULL,
+    time BIGINT NOT NULL,
     vessel_name TEXT,
     ship_type INTEGER,
     call_sign TEXT,
