+++ aisdb/aisdb_sql/timescale_createtable_dynamic.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS ais_global_dynamic
 (
-    mmsi          INTEGER NOT NULL,
-    time          INTEGER NOT NULL,
+    mmsi          BIGINT NOT NULL,
+    time          BIGINT NOT NULL,
     longitude     REAL NOT NULL,
     latitude      REAL NOT NULL,
     rot           REAL,
